feat: support windows dev env
Works in Git Bash and PowerShell in Windows 10
