+++ pkg/rtc/rtc_windows.go
@@ -1,0 +1,7 @@
+//go:build windows
+// +build windows
+
+package rtc
+
+func checkUDPReadBuffer() {
+}
